refactor persistencyy in core detector component
