+++ src/detectmatelibrary/common/detector.py
@@ -6,38 +6,16 @@ from detectmatelibrary.utils.aux import get_timestamp
 from detectmatelibrary.utils import persistency
 from detectmatelibrary.common.persist import init_persistency
 
-import os
-
-from pydantic import BaseModel, ConfigDict, Field
-
 from detectmatelibrary.schemas import ParserSchema, DetectorSchema
 
 from typing_extensions import override
 from typing import Dict, List, Optional, Any, cast
 
+from detectmatelibrary.utils.persistency.component_interfaces import PersistConfig
 from detectmatelibrary.utils.time_format_handler import TimeFormatHandler
-from detectmatelibrary.tools.logging import logger
 
 
 _time_handler = TimeFormatHandler()
-
-
-class PersistConfig(BaseModel):
-    model_config = ConfigDict(extra="forbid")
-
-    # Default honors systemd's $STATE_DIRECTORY (set by StateDirectory= in the
-    # unit file) so services persist to /var/lib/<dir> with no explicit path=.
-    # Falls back to CWD-relative ./state outside systemd. Explicit path= wins.
-    path: str = Field(
-        default_factory=lambda: next(
-            (p for p in os.environ.get("STATE_DIRECTORY", "").split(":") if p.strip()),
-            "./state",
-        )
-    )
-    interval_seconds: int = 300
-    events_until_save: int | None = None
-    auto_load: bool = False
-    storage_options: dict[str, Any] = {}
 
 
 def _extract_timestamp(
@@ -55,102 +33,6 @@ def _extract_logIDs(
         input_ = [input_]
 
     return [str(i["logID"]) for i in input_]
-
-
-def get_configured_variables(
-        input_: ParserSchema,
-        log_variables: EventsConfig | dict[str, Any],
-) -> Dict[str, Any]:
-    """Extract variables from input based on what's defined in the config.
-
-    Args:
-        input_: Parser schema containing variables and logFormatVariables
-        log_variables: Config specifying which variables to extract per EventID
-
-    Returns:
-        Dict mapping variable names to their values from the input
-    """
-    event_id = input_["EventID"]
-    result: Dict[str, Any] = {}
-
-    # Get the config for this event
-    event_config = log_variables[event_id] if event_id in log_variables else None
-    if event_config is None:
-        return result
-
-    # Extract template variables by position
-    if hasattr(event_config, "variables"):
-        for pos, var in event_config.variables.items():
-            if isinstance(pos, int) and pos < len(input_["variables"]):
-                result[var.name] = input_["variables"][pos]
-
-    # Extract header/log format variables by name
-    if hasattr(event_config, "header_variables"):
-        for name in event_config.header_variables:
-            if name in input_["logFormatVariables"]:
-                result[name] = input_["logFormatVariables"][name]
-
-    return result
-
-
-def get_global_variables(
-        input_: ParserSchema,
-        global_instances: Dict[str, _EventInstance],
-) -> Dict[str, Any]:
-    """Extract header variables from event-ID-independent instances.
-
-    Args:
-        input_: Parser schema containing logFormatVariables
-        global_instances: Dict of instance_name -> _EventInstance configs
-
-    Returns:
-        Dict mapping variable names to their values from the input
-    """
-    result: Dict[str, Any] = {}
-    for instance in global_instances.values():
-        for name in instance.header_variables:
-            if name in input_["logFormatVariables"]:
-                result[name] = input_["logFormatVariables"][name]
-    return result
-
-
-def validate_config_coverage(
-        detector_name: str,
-        config_events: EventsConfig | dict[str, Any],
-        event_persistency: persistency.EventPersistency,
-) -> None:
-    """Log warnings when configured EventIDs or variables have no training
-    data.
-
-    Args:
-        detector_name: Name of the detector (used in warning messages).
-        config_events: The detector's events configuration.
-        persistency: The persistency object populated during training.
-    """
-    config_ids = (
-        config_events.events.keys()
-        if isinstance(config_events, EventsConfig)
-        else config_events.keys()
-    )
-    if not config_ids:
-        return
-
-    events_seen = event_persistency.get_events_seen()
-    events_with_data = set(event_persistency.get_events_data().keys())
-
-    for event_id in config_ids:
-        if event_id not in events_seen:
-            logger.warning(
-                f"[{detector_name}] EventID {event_id!r} is configured but was "
-                "never observed in training data. Verify that EventIDs in your "
-                "config match those produced by the parser."
-            )
-        elif event_id not in events_with_data:
-            logger.warning(
-                f"[{detector_name}] EventID {event_id!r} was observed in training "
-                "data but no configured variables were extracted. Verify that "
-                "variable names/positions in your config match those in the data."
-            )
 
 
 class CoreDetectorConfig(CoreConfig):
